requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
tensorflow<=2.15.0
keras==2.15.0
pandas<=2.1.3
scikit-learn<=1.3.2
opencv-python==4.8.1.78
matplotlib<=3.8.0
seaborn<=0.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1

